requirements.txt
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dotenv==0.19.1
Werkzeug==2.2.3
Flask-Login==0.6.2
mysql-connector-python==8.0.26
click==7.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
WTForms==2.3.3
Flask-WTF==1.1.1